refactor(events-import-api): some restructuring
