+++ libs/events/import/api/src/index.ts
@@ -1,2 +1,4 @@
 export * from './lib/events-import.module'
 export * from './lib/import/agenda-basel.service'
+export * from './lib/import/events-import.service'
+export * from './lib/import/kulturagenda-parser'
