+++ libs/image/api/src/index.ts
@@ -1,3 +1,5 @@
 export * from './lib/image-upload'
 export * from './lib/media-adapter'
 export * from './lib/media-adapter.service'
+export * from './lib/image-fetcher/image-fetcher.module'
+export * from './lib/image-fetcher/image-fetcher.service'
