+++ libs/image/api/src/lib/image-fetcher/image-fetcher.service.ts
@@ -1,0 +1,26 @@
+import {HttpService} from '@nestjs/axios'
+import {firstValueFrom} from 'rxjs'
+import {ArrayBufferUpload} from '../media-adapter'
+
+export class ImageFetcherService {
+  constructor(private httpService: HttpService) {}
+
+  async fetch(url: string, filename = 'transformed-image.jpg'): Promise<ArrayBufferUpload> {
+    try {
+      const {data: arrayBuffer, headers} = await firstValueFrom(
+        this.httpService.get(url, {responseType: 'arraybuffer'})
+      )
+      const mimetype = headers['content-type']
+
+      const arrayBufferUpload = {
+        filename,
+        mimetype,
+        arrayBuffer
+      }
+
+      return arrayBufferUpload
+    } catch (error) {
+      throw Error('Error fetching and transforming image')
+    }
+  }
+}
